pyproject.toml
[project]
name = "pdf-debug"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bson>=0.5.10",
    "numpy>=2.3.2",
    "pillow>=11.3.0",
    "pymongo>=4.15.0",
    "pymupdf>=1.26.4",
    "pypdf>=6.0.0",
]
